handle wrong value problems in data_manager

+++ data_manager.py
@@ -1,5 +1,3 @@
-from sqlalchemy.util import method_is_overridden
-
 from models import db, User, Movie
 
 
@@ -20,21 +18,29 @@ class DataManager():
         return user.movies
 
     def add_movie(self, movie):
+        if not movie:
+            return
         db.session.add(movie)
         db.session.commit()
 
     def update_movie(self, movie_id, new_title):
         movie_to_update = Movie.query.get(movie_id)
+        if not movie_to_update:
+            return
         movie_to_update.title = new_title
         db.session.commit()
 
     def delete_movie(self, movie_id):
         movie_to_delete = Movie.query.get(movie_id)
+        if not movie_to_delete:
+            return
         db.session.delete(movie_to_delete)
         db.session.commit()
 
     def delete_user(self, user_id):
         user_to_delete = User.query.get(user_id)
+        if not user_to_delete:
+            return
         db.session.delete(user_to_delete)
         db.session.commit()
